requirements.txt
PyYAML>=6.0
keybert>=0.9.0
sentence-transformers>=5.1.0
spacy>=3.8.0
textblob>=0.19.0
torch>=2.8.0
transformers>=4.55.0
vaderSentiment>=3.3.2
yake>=0.6.0
